# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * iPad AI Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
  tool: "tool",
} as const;

export type ChatMessageToolCallsItem = {
  tool: string;
  input: string;
  output: string;
};

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  timestamp: string;
  toolCalls?: ChatMessageToolCallsItem[];
  thinking?: string;
}

export interface ChatRequest {
  message: string;
  includeTools?: boolean;
}

export type ChatResponseToolsUsedItem = {
  tool: string;
  input: string;
  output: string;
};

export interface ChatResponse {
  reply: string;
  thinking?: string;
  toolsUsed?: ChatResponseToolsUsedItem[];
  messageId: string;
}

export interface ChatHistory {
  messages: ChatMessage[];
}

export interface TerminalCommandRequest {
  command: string;
  timeout?: number;
}

export interface TerminalCommandResponse {
  stdout: string;
  stderr: string;
  exitCode: number;
  success: boolean;
  duration: number;
}

export type TerminalStatusMode =
  (typeof TerminalStatusMode)[keyof typeof TerminalStatusMode];

export const TerminalStatusMode = {
  ssh: "ssh",
  local: "local",
  disconnected: "disconnected",
} as const;

export interface TerminalStatus {
  connected: boolean;
  host?: string;
  port?: number;
  mode: TerminalStatusMode;
  lastCommand?: string;
}

export type BrowserNavigateRequestSource =
  (typeof BrowserNavigateRequestSource)[keyof typeof BrowserNavigateRequestSource];

export const BrowserNavigateRequestSource = {
  shortcuts: "shortcuts",
  playwright: "playwright",
  manual: "manual",
} as const;

export interface BrowserNavigateRequest {
  url: string;
  source?: BrowserNavigateRequestSource;
}

export type BrowserActionRequestAction =
  (typeof BrowserActionRequestAction)[keyof typeof BrowserActionRequestAction];

export const BrowserActionRequestAction = {
  click: "click",
  fill: "fill",
  scrape: "scrape",
  screenshot: "screenshot",
  evaluate: "evaluate",
} as const;

export interface BrowserActionRequest {
  action: BrowserActionRequestAction;
  selector?: string;
  value?: string;
  script?: string;
}

export interface BrowserActionResponse {
  success: boolean;
  result: string;
  url?: string;
  title?: string;
}

export interface BrowserScreenshotResponse {
  success: boolean;
  imageData: string;
  url?: string;
}

export interface BrowserStatus {
  active: boolean;
  currentUrl?: string;
  currentTitle?: string;
  lastAction?: string;
  source?: string;
}

export interface ShortcutsWebhookRequest {
  action: string;
  url?: string;
  pageContent?: string;
  pageTitle?: string;
  screenshotBase64?: string;
  timestamp?: string;
}

export interface ShortcutsWebhookResponse {
  received: boolean;
  nextAction?: string;
  data?: string;
}

export type FileItemType = (typeof FileItemType)[keyof typeof FileItemType];

export const FileItemType = {
  file: "file",
  directory: "directory",
} as const;

export interface FileItem {
  name: string;
  size: number;
  type: FileItemType;
  modified: string;
  mimeType?: string;
}

export interface FileListResponse {
  files: FileItem[];
  path: string;
}

export interface FileUploadResponse {
  success: boolean;
  filename: string;
  size: number;
  path: string;
}

export interface ImageAnalysisRequest {
  /** Base64 encoded image */
  imageData: string;
  prompt?: string;
  mimeType?: string;
}

export interface ImageAnalysisResponse {
  description: string;
  objects?: string[];
  confidence?: number;
  model: string;
}

export interface AppSettings {
  groqApiKey?: string;
  llmModel: string;
  sshHost?: string;
  sshPort?: number;
  sshUsername?: string;
  enableCamera: boolean;
  enableFileAccess: boolean;
  enableTerminal: boolean;
  enableBrowser: boolean;
  maxLoops?: number;
  agentTimeout?: number;
}

export interface AppSettingsUpdate {
  groqApiKey?: string;
  llmModel?: string;
  sshHost?: string;
  sshPort?: number;
  sshUsername?: string;
  sshPassword?: string;
  enableCamera?: boolean;
  enableFileAccess?: boolean;
  enableTerminal?: boolean;
  enableBrowser?: boolean;
  maxLoops?: number;
  agentTimeout?: number;
}

export type SetupGuideStepsItem = {
  step: number;
  title: string;
  description: string;
  command?: string;
};

export interface SetupGuide {
  steps: SetupGuideStepsItem[];
  qrCodeUrl?: string;
  shortcutsUrl?: string;
  permissions: string[];
  webhookUrl: string;
}

export type ListFilesParams = {
  path?: string;
};

export type UploadFileBody = {
  file?: Blob;
  path?: string;
};

export type DeleteFileBody = {
  filename: string;
};