# lib/api-zod/src/generated/types/appSettingsUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * iPad AI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AppSettingsUpdate {
  groqApiKey?: string;
  llmModel?: string;
  sshHost?: string;
  sshPort?: number;
  sshUsername?: string;
  sshPassword?: string;
  enableCamera?: boolean;
  enableFileAccess?: boolean;
  enableTerminal?: boolean;
  enableBrowser?: boolean;
  maxLoops?: number;
  agentTimeout?: number;
}