# lib/api-zod/src/generated/types/fileItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * iPad AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { FileItemType } from "./fileItemType";

export interface FileItem {
  name: string;
  size: number;
  type: FileItemType;
  modified: string;
  mimeType?: string;
}