# lib/api-zod/src/generated/types/fileListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * iPad AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { FileItem } from "./fileItem";

export interface FileListResponse {
  files: FileItem[];
  path: string;
}