# lib/api-zod/src/generated/types/imageAnalysisResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * iPad AI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ImageAnalysisResponse {
  description: string;
  objects?: string[];
  confidence?: number;
  model: string;
}